feat: add session start limit to non-sharded client
