+++ disnake/client.py
@@ -30,6 +30,7 @@ import logging
 import signal
 import sys
 import traceback
+from datetime import datetime, timedelta
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -94,7 +95,7 @@ if TYPE_CHECKING:
     from .member import Member
     from .message import Message
     from .role import Role
-    from .shard import SessionStartLimit
+    from .types.gateway import SessionStartLimit as SessionStartLimitPayload
     from .voice_client import VoiceProtocol
 
 
@@ -139,6 +140,50 @@ def _cleanup_loop(loop: asyncio.AbstractEventLoop) -> None:
     finally:
         _log.info("Closing the event loop.")
         loop.close()
+
+
+class SessionStartLimit:
+    """A class that contains information about the current session start limit.
+
+    .. versionadded:: 2.5
+
+    Attributes
+    ----------
+    total: :class:`int`
+        The total number of allowed session starts.
+    remaining: :class:`int`
+        The remaining number of allowed session starts.
+    reset_after: :class:`int`
+        The number of milliseconds after which the :attr:`.remaining` limit resets,
+        relative to when the client connected.
+        See also :attr:`reset_time`.
+    max_concurrency: :class:`int`
+        The number of allowed ``IDENTIFY`` requests per 5 seconds.
+    reset_time: :class:`datetime.datetime`
+        The approximate time at which which the :attr:`.remaining` limit resets.
+    """
+
+    __slots__: Tuple[str, ...] = (
+        "total",
+        "remaining",
+        "reset_after",
+        "max_concurrency",
+        "reset_time",
+    )
+
+    def __init__(self, data: SessionStartLimitPayload):
+        self.total: int = data["total"]
+        self.remaining: int = data["remaining"]
+        self.reset_after: int = data["reset_after"]
+        self.max_concurrency: int = data["max_concurrency"]
+
+        self.reset_time: datetime = utils.utcnow() + timedelta(milliseconds=self.reset_after)
+
+    def __repr__(self):
+        return (
+            f"<SessionStartLimit total={self.total!r} remaining={self.remaining!r} "
+            f"reset_after={self.reset_after!r} max_concurrency={self.max_concurrency!r} reset_time={self.reset_time}"
+        )
 
 
 class Client:
@@ -267,8 +312,8 @@ class Client:
         Whether to enable asyncio debugging when the client starts.
         Defaults to False.
     session_start_limit: Optional[:class:`SessionStartLimit`]
-        Information about the current session start limit,
-        only set in sharded clients/bots.
+        Information about the current session start limit.
+        Only available after initiating the connection.
 
         .. versionadded:: 2.5
     """
@@ -692,11 +737,16 @@ class Client:
         ConnectionClosed
             The websocket connection has been terminated.
         """
-        backoff = ExponentialBackoff()
+        _, gateway, session_start_limit = await self.http.get_bot_gateway()
+        self.session_start_limit = SessionStartLimit(session_start_limit)
+
         ws_params = {
             "initial": True,
             "shard_id": self.shard_id,
+            "gateway": gateway,
         }
+
+        backoff = ExponentialBackoff()
         while not self.is_closed():
             try:
                 coro = DiscordWebSocket.from_client(self, **ws_params)
